lambda/requirements.txt
boto3
python-docx
PyPDF2
textract